RichsFirstAttemptAtACommit
This is Rich attempting his first basic pull request. In these edits,
I've added the 'outlines' of the new report pages we previously
discussed. This includes 3 new views, edits to the URL.py file, and 3
new templates. More to come obviously, but I just wanted to say and see
if I can do this correctly. I've run this on my local machine and all is
working w/o errors and i'm able to see and interact with my new web
pages as expected.