Rich_1st_version_of_CSV_Reporting_Form_and_small_mobile_app_fixes
This is Rich's first take at a functional django-based web form that
allows users to request a CSV file be generated that facilitates
integration into an existing database. From the reports section of the
web app, it allows users to pull survey data from a single appointment,
or all appointments for a given year. One ajax call is used to
dynamically update the "available appointments" dropdown depending on
the user's selection of the "count year". I still need to improve the
"data cleansing/validation" on the backend to improve security. CSV file
is dynamically generated after user submits form. ....In the process of
making updates I found a few bugs in the mobile survey app's survey
experience that was leading to inaccurate data being reported, so this
pull request includes some of those changes too.