April24_Update
Updated pause and timer logic to store the expected end time in local
storage vs the elapsed time. This allows the timer to always reflect the
correct remaining time - even if the browser window loses focus. In
doing so I updated the pause logic and added the "total time spent in
current pause" value to the screen. I also fixed the "reset" admin
function so it truly resets all data with an appointment. Finally, I
added a link to website where users could go signup for more
appointments, to the upcoming table.